air-quality-monitor/requirements.txt
streamlit==1.28.1
pandas==2.1.3
plotly==5.17.0
requests==2.31.0
numpy==1.24.3
scikit-learn==1.3.2
statsmodels==0.14.0
prophet==1.1.4
folium==0.14.0
streamlit-folium==0.13.0
python-dotenv==1.0.0
streamlit-option-menu==0.3.6
streamlit-autorefresh==1.3.1
